% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_functions.R
\name{anglefun}
\alias{anglefun}
\title{Title}
\usage{
anglefun(xx, yy, bearing = TRUE, as.deg = FALSE)
}
\arguments{
\item{xx}{TODO}

\item{yy}{TODO}

\item{bearing}{TODO}

\item{as.deg}{TODO}
}
\value{
TODO
}
\description{
Title
}
