% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_corrupted_mc.R
\name{check_corrupted_mc}
\alias{check_corrupted_mc}
\title{Multicore version of check_corrupted function which uses the  showinf tool from the bftools suite (provided by BIO-LOCI)}
\usage{
check_corrupted_mc(to.data, raw.data.folder, path_showinf)
}
\arguments{
\item{to.data}{path to the working directory}

\item{raw.data.folder}{directory with the raw video files}

\item{path_showinf}{path to the showinf executable}
}
\description{
The function will output the names of uncorruped and corrupted files, which would make BEMOVI bug; the files get
transferred into a "bad files" directory in the raw.video.folder and should be either repaired or ignored for 
further analysis.
}
