% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tools_folder.R
\name{check_tools_folder}
\alias{check_tools_folder}
\title{Function to check if all dependencies are present, and if not download them}
\usage{
check_tools_folder(tools.path)
}
\arguments{
\item{tools.path}{path to the folder where the dependencies (fiji, bftools and ffmpeg) should be located}
}
\value{
returns nothing (NULL)
}
\description{
Function calls ImageJ software and its ParticleAnalyzer function to extract for each frame of the video
several morphological descriptors and the X- and Y-coordinates of all moving particles. All videos in the raw.video.folder are analyses, separately.
}
