% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_functions.R
\name{convert_cxd_to_avi}
\alias{convert_cxd_to_avi}
\title{Function to convert the video files to .avi format using lossless conversion}
\usage{
convert_cxd_to_avi(
  cxd_file,
  avi_dir,
  compression_level = 5,
  ffmpeg = par_ffmpeg(),
  bfconvert = par_bfconvert(),
  showinf = par_showinf(),
  mc.cores = par_mc.cores()
)
}
\arguments{
\item{cxd_file}{one or more \code{cxd} file to be converted or a directory with \code{.cxd} files.}

\item{avi_dir}{directory for the converted \code{cxd} files and the metadata files}

\item{compression_level}{compression level - defaults to 5}

\item{ffmpeg}{execuable ffmpeg. May have to be including path.}

\item{bfconvert}{executable bfconvert from bftools. May have to be including path.}

\item{showinf}{executable showinf from bftools. May have to be including path.}

\item{mc.cores}{Number of cores to be used when more than one xd file is given.}
}
\value{
returns nothing (NULL)
}
\description{
Function uses bftools to convert \code{cxd} files to \code{avi} and ffmpeg to 
compress these to lossles avi
}
