% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_avi.R
\name{convert_to_avi}
\alias{convert_to_avi}
\title{Function to convert the video files to .avi format using lossless conversion}
\usage{
convert_to_avi(
  to.data,
  raw.video.folder,
  raw.avi.folder,
  metadata.folder,
  tools.path,
  fps,
  video.format
)
}
\arguments{
\item{to.data}{path to the working directory}

\item{raw.video.folder}{directory with the raw video files}

\item{raw.avi.folder}{directory to store the converted avi files}

\item{metadata.folder}{directory to store the metadata from the .cxd files}

\item{tools.path}{directory where the tools for bemovi (fiji, ffmpeg and bftools) are stored}

\item{fps}{framerate of the videos}

\item{video.format}{extension of the video files (e.g. cxd, avi, mp4, mov)}
}
\value{
returns nothing (NULL)
}
\description{
Function uses bftools to convert cxd files and ffmpeg to convert other files and compress to lossles avi
}
