% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Folder_structure.R
\name{Create_folder_structure}
\alias{Create_folder_structure}
\title{Create folder structure based on parameter}
\usage{
Create_folder_structure(to.data = par_to.data(),
  video.description.folder = par_video.description.folder(),
  raw.video.folder = par_raw.video.folder(),
  particle.data.folder = par_particle.data.folder(),
  trajectory.data.folder = par_trajectory.data.folder(),
  temp.overlay.folder = par_temp.overlay.folder(),
  overlay.folder = par_overlay.folder(),
  merged.data.folder = par_merged.data.folder(),
  ijmacs.folder = par_ijmacs.folder())
}
\arguments{
\item{to.data}{path to the working directory}

\item{video.description.folder}{directory containing the video description file}

\item{raw.video.folder}{directory with the raw video files}

\item{particle.data.folder}{directory to which the data is saved as a text file}

\item{trajectory.data.folder}{directory containing the global trajectory data}

\item{temp.overlay.folder}{temporary directory to save the overlay created with R}

\item{overlay.folder}{directory where the overlay videos are saved}

\item{merged.data.folder}{directory where the global database is saved}

\item{ijmacs.folder}{directory for the macro to for ImageJ}
}
\value{
invisibly \code{TRUE}
}
\description{
Create folder structure based on parameter
}
