% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_overlay_videos_ffmpeg.R
\name{create_overlays_ffmpeg}
\alias{create_overlays_ffmpeg}
\title{Function to make overlays using the subtitle files,}
\usage{
create_overlays_ffmpeg(
  to.data,
  traj.data,
  raw.avi.folder,
  temp.overlay.folder,
  overlay.folder,
  fps,
  vid.length,
  width,
  height
)
}
\arguments{
\item{to.data}{path to the working directory}

\item{traj.data}{filtered trajectory file}

\item{raw.avi.folder}{path to the folder containing the converted and compressed .avi files}

\item{temp.overlay.folder}{temporary directory to save the overlay subtitles (.ssa files)}

\item{overlay.folder}{directory where the overlay videos are saved}

\item{fps}{framerate in the original video}

\item{vid.length}{duration in seconds of the video}

\item{width}{width in pixels of the original videos}

\item{height}{height in pixels of the original video}
}
\value{
returns nothing (NULL)
}
\description{
Function creates subtitle comands for every cell in every frame of the videos, using the x/y coordinates.
Then ffmpeg is called to burn the overlay subtitles on, and save a compressed video.
}
