% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_overlay_videos_parallel.R
\name{create_overlays_parallel}
\alias{create_overlays_parallel}
\title{Parallellized version of the function to create a new video with the extracted trajectories (filtered or raw) overlayed onto the original video}
\usage{
create_overlays_parallel(
  process_ID,
  video.files.df,
  traj.data,
  to.data,
  merged.data.folder,
  raw.video.folder,
  temp.overlay.folder,
  overlay.folder,
  width,
  height,
  difference.lag,
  type = "traj",
  predict_spec = F,
  contrast.enhancement = 0,
  IJ.path,
  memory = 512
)
}
\arguments{
\item{process_ID}{numeric variable, containing the identifier for the parallel process}

\item{video.files.df}{Dataframe containing the video files in the raw data, as well as the process ID for the parallel analysis}

\item{traj.data}{dataframe with the information on morphology and movement for each frame (either "trajectoty.data" or "traj.data.filtered")}

\item{to.data}{path to the working directory}

\item{merged.data.folder}{directory where the global database is saved}

\item{raw.video.folder}{directory with the raw video files}

\item{temp.overlay.folder}{temporary directory to save the overlay created with R}

\item{overlay.folder}{directory where the overlay videos are saved}

\item{width}{width of the raw video}

\item{height}{height of the raw video}

\item{difference.lag}{numeric value specifying the offset between two video frames to compute the difference image}

\item{type}{string indicating the visualization type (i.e. 'label' or 'traj'): either the overlay
is showing the trajectory ID and outlines the detected particle (type='label') or the whole trajectory
remains plotted (type='traj').}

\item{predict_spec}{logical If TRUE, the Master.RData file must have a column called predict_spec, indicating the species to which the trajectory belongs}

\item{contrast.enhancement}{numeric value to increase the contrast of the original video}

\item{IJ.path}{path to ImageJ folder, containing the 'ij.jar' executable}

\item{memory}{numeric value specifying the amount of memory available to ImageJ (defaults to 512)}
}
\description{
A function to parallelize the overlay generation
}
