% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_videos.R
\name{crop_videos}
\alias{crop_videos}
\title{Function to CROP VIDEOS THAT NEED TO BE CROPPED (by Fanny Gonguet)}
\usage{
crop_videos(
  to.data,
  particle.data.folder,
  max.weight,
  coordinates.threshold,
  width,
  height
)
}
\arguments{
\item{to.data}{path to the working directory}

\item{particle.data.folder}{directory in which text file data describing particules morphology and coordiantes are stored}

\item{max.weight}{variable to set as weight from wich text file data should be cropped (Mb)}

\item{coordinates.threshold}{variable to set as the minimum coordinates from wich to crop the particles coordinates (pixels)}

\item{width}{width in pixels of the original videos}

\item{height}{height in pixels of the original video}
}
\value{
returns nothing (NULL)
}
\description{
Function to reduce the time of analyses for too heavy documents by cropping particles coordinates. Only them whithin the cropped frame will continue the analyse.
}
