\docType{package}
\name{fRanco}
\alias{fRanco}
\alias{fRanco-package}
\title{fRanco - an R package to quantify abundance, traits and behaviour of individuals in aquatic microbial communities}
\description{
fRanco provides functions to segment videos of aquatic
microcosms, identify individuals and their traits
(morphology and movement) and measure their abundances
}

