% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxd_functions.R
\name{get_delays_cxd}
\alias{get_delays_cxd}
\title{Extract delays from the \code{.cxd} file}
\usage{
get_delays_cxd(file)
}
\arguments{
\item{file}{file name, of the \code{.cxd} file}
}
\value{
a vector of the length of number of frames -1 specifying the delay between the images taken.
}
\description{
This function reads the metadata from the \code{.cxd} file and returns the 
delays between the images were taken.
}
