% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxd_functions.R
\name{get_duration_ffmpeg}
\alias{get_duration_ffmpeg}
\title{Extract duration in seconds using ffmpeg}
\usage{
get_duration_ffmpeg(file)
}
\arguments{
\item{file}{file name, of the video file}
}
\value{
the fps (frames per second) of the video
}
\description{
The duration is based on using the output of \code{ffmpeg} and
should therefore work for all ffmpeg supported video formats.
}
