% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_functions.R
\name{get_width_avi}
\alias{get_width_avi}
\title{Extract width in pixels using ffmpeg}
\usage{
get_width_avi(file, mc.cores = par_mc.cores())
}
\arguments{
\item{file}{file name, of the video file}

\item{mc.cores}{Number of cores to be used when more than one xd file is given.}
}
\value{
named vector with the width (in pixels) of the video(s) and the names are the file names
}
\description{
The width is based on using the output of \code{ffmpeg} and
should therefore work for all ffmpeg supported video formats,
but is only implemented for \code{.avi} files.
}
