% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ij.bin.R
\name{ij.bin}
\alias{ij.bin}
\title{Return ImegeJ executable based on platform. Darwin, Windows and Linux are supported}
\usage{
ij.bin()
}
\value{
full path of ewxecutable of ImageJ
}
\description{
Return ImegeJ executable based on platform. Darwin, Windows and Linux are supported
}
\examples{
ij.bin()

}
