% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{load_parameter}
\alias{load_parameter}
\title{Load parameter from \code{file}}
\usage{
load_parameter(file = "parameter.yaml")
}
\arguments{
\item{file}{name of parameter file}
}
\value{
invisibly TRUE
}
\description{
Load parameter from \code{file}
}
