% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Function to create global database containg morphology, trajectory, and video (e.g., experimental) information}
\usage{
merge_data(to.data = par_to.data(),
  particle.data.folder = par_particle.data.folder(),
  trajectory.data.folder = par_trajectory.data.folder(),
  video.description.folder = par_video.description.folder(),
  video.description.file = par_video.description.file(),
  merged.data.folder = par_merged.data.folder())
}
\arguments{
\item{to.data}{path to the working directory}

\item{particle.data.folder}{directory containing the global morphology data}

\item{trajectory.data.folder}{directory containing the global trajectory data}

\item{video.description.folder}{directory containing the video description file}

\item{video.description.file}{name of the video description file}

\item{merged.data.folder}{directory where the global database is saved}
}
\value{
saves the global database Master.rds to the merged.data.folder
}
\description{
Merges the morphology data, the trajectory data with the video descriptions, which can / should contain 
the information on sampling units, video date and time, treatments and replicate etc. The files are merged by the use of the
video file names. For the exact meaning of each of the columns, please refer to the locate_and_measure_particles() and link_particles() functions.
}
