\name{Organise_particle_data}
\alias{Organise_particle_data}
\title{Function to merge the morphology and data on X- and Y-coordinates into one file for further processing}
\usage{
Organise_particle_data(to.data, particle.data.folder)
}
\arguments{
  \item{path}{Path to the text files containing the
  morphology and coordinate data}
}
\description{
This function merges the files containing morphology and
coordinates (one for each video) into large dataset, and
saves it to the directory where the single files are
located
}

