% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_bfconvert}
\alias{par_bfconvert}
\title{parameter executable bfconvert of the bftools commandline tools}
\usage{
par_bfconvert(value)
}
\description{
parameter executable bfconvert of the bftools commandline tools
}
