% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_calc_crop}
\alias{par_calc_crop}
\title{If cropped, uncropped or both d=should be calculated}
\usage{
par_calc_crop(value)
}
\description{
If cropped, uncropped or both d=should be calculated
}
\details{
Can contain the following values:
  \code{cropped}: calculate cropped only;
  \code{uncroped}; calculate uncropped only;
  \code{both}: calculate both
}
