% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_classifier_constant}
\alias{par_classifier_constant}
\title{Name of classifier file for constant temperature treatment}
\usage{
par_classifier_constant(value)
}
\description{
Name of classifier file for constant temperature treatment
}
