% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_classifier_increasing}
\alias{par_classifier_increasing}
\title{Name of classifier file for increasing temperature treatment}
\usage{
par_classifier_increasing(value)
}
\description{
Name of classifier file for increasing temperature treatment
}
