% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_crop_pixels}
\alias{par_crop_pixels}
\title{Cropping of identified and measured particles.}
\usage{
par_crop_pixels(value)
}
\description{
Cropping of identified and measured particles.
}
\details{
This function returns a list with four elements, namely \code{xmin}, \code{xmax}, \code{ymin} and \code{ymax}.
  To set this parameter \bold{all need to be set in a named list!}
  If a value is \code{NULL} (\code{~} in the parameter file), it is assumed to be the maximum or minimum value.
}
