% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_mc.cores}
\alias{par_mc.cores}
\title{Number of cores to be used for parallel processing. Defaults to 1, i.e. no parallel processing}
\usage{
par_mc.cores(value)
}
\description{
Number of cores to be used for parallel processing. Defaults to 1, i.e. no parallel processing
}
