% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_merged.dataunfiltered..folder}
\alias{par_merged.dataunfiltered..folder}
\alias{par_merged.data.unfiltered.folder}
\title{parameter merged.data.unfiltered.folder}
\usage{
par_merged.data.unfiltered.folder(value)
}
\description{
parameter merged.data.unfiltered.folder
}
