% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_showinf}
\alias{par_showinf}
\title{parameter executable showinf of the bftools commandline tools}
\usage{
par_showinf(value)
}
\description{
parameter executable showinf of the bftools commandline tools
}
