% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{par_template}
\alias{par_template}
\title{Template function to assign value to parameter in the package wide cache}
\usage{
par_template(value)
}
\arguments{
\item{value}{if missing, the value of the parameter will be returned, \code{NULL} if the parameter does not exist; if specified, the parameter will be set to the value}
}
\value{
the (new) value of the argument
}
\description{
assign the function to a new cvariable and the name of the function woll be used for the parameter name. e.g:
\itemize{
   \item{fps <- par_template}
}
}
