% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Locate_and_measure_particles_parallel.R
\name{parallel_locate_and_measure}
\alias{parallel_locate_and_measure}
\title{Function to extract morphological measurements and X- and Y-coordinates for moving particles, in a wrapper for parallelized analyses}
\usage{
parallel_locate_and_measure(
  process_ID,
  to.data,
  raw.avi.folder,
  particle.data.folder,
  difference.lag,
  min_size,
  max_size,
  thresholds,
  tools.path,
  memory,
  video.files.df
)
}
\arguments{
\item{process_ID}{numeric variable, containing the identifier for the parallel process}

\item{to.data}{path to the working directory}

\item{raw.avi.folder}{directory with the compressed avi files}

\item{particle.data.folder}{directory to which the data is saved as a text file}

\item{difference.lag}{numeric value specifying the offset between two video frames to 
compute the difference image. If 0, then no differencing applied.}

\item{min_size}{minimum size for detection of particles}

\item{max_size}{maximum size for detection of particles}

\item{thresholds}{vector containing the min and max threshold values (defaults to c(10,255))}

\item{tools.path}{Path containing the bemovi dependencies}

\item{memory}{numeric value specifying the amount of memory available to ImageJ}

\item{video.files.df}{Dataframe containing the video files in the raw data, as well as the process ID for the parallel analysis}
}
\value{
saves the output of the ParticleAnalyzer function of ImageJ as a text file in the output directory and then assembles the data into a single database 
called 'particle.RData'. This data.frame contains information about the following properties: the area (transversal cut), 
the mean, minimum and maximum of the grey value, the perimeter, width, length and angle with the dominant-axis of a fitted ellipse, and finally shape parameters such as
circularity, aspect ratio, roundness and solidity. For details of the morphological output, please refer to http://rsbweb.nih.gov/ij/docs/guide/146-30.html
}
\description{
Function calls ImageJ software and its ParticleAnalyzer function to extract for each frame of the video
several morphological descriptors and the X- and Y-coordinates of all moving particles. All videos in the raw.video.folder are analyses, separately.
To parallellize this process, we create temporary folders cotaining the subset of videos to be analyzed, as well as containing a temporary
folder with an ImageJ/Java copy
}
