% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{print_parameter}
\alias{print_parameter}
\title{Print the bemovi parameter}
\usage{
print_parameter(print_as_yaml = TRUE, echo = TRUE)
}
\arguments{
\item{print_as_yaml}{Print in yaml formated text; \code{~} stands for NULL}

\item{should}{anything be printed (\code{TRUE}) or just a list returned (\code{FALSE})}
}
\value{
invisible returns list of parameter for further processing
}
\description{
Print the bemovi parameter
}
