% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{save_parameter}
\alias{save_parameter}
\title{Save parameter into \code{.yaml} file}
\usage{
save_parameter(file = "parameter.yaml")
}
\arguments{
\item{file}{name of parameter file}
}
\value{
invisibly \code{TRUE}
}
\description{
Save parameter into \code{.yaml} file
}
