% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch_trajectory.R
\name{stitch}
\alias{stitch}
\title{Function to verify and stitch the trajectories of Link_particles}
\usage{
stitch(to.data, merged.data.folder, stitch.data.folder, accuracy = 0.5,
  morphological_parameter, search_area = 20)
}
\description{
in previous trajectory.data, we could find: one individuals with multiple trajectories or one trajectory for multiple individuals
the goal here is to have one trajectory for one individuals on the script master.worker, at a new folder on the intermediate directories
stitch.data.folder <- "6 - stitch data/"
}
\details{
STEP 1: identify a change within trajectory
STEP 2: create a new_trajectory 
STEP 3: summarise morphological information to search similar individuals
STEP 4: search individuals in other trajectories
STEP 5: replace new trajectories to use them for overlays
}
\author{
Aurélie Garnier
}
