% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_populations.R
\name{summarize_populations}
\alias{summarize_populations}
\title{Function to calculate densities and summarize morphology and movement at the population (sample) level}
\usage{
summarize_populations(
  traj.data,
  sum.data,
  write = FALSE,
  to.data,
  merged.data.folder,
  video.description.folder,
  video.description.file,
  total_frames,
  coordinates.threshold,
  width,
  height
)
}
\arguments{
\item{traj.data}{dataframe with the information on morphology and movement for each frame (either "trajectoty.data" or "trajectory.data.filtered")}

\item{sum.data}{data.table with the aggregated morphology and movement information for each trajectory ("morph_mvt")}

\item{write}{logical argument to indicate whether aggregated information should be saved to disk (file name: "Population_Data.RData")}

\item{to.data}{path to the working directory}

\item{merged.data.folder}{directory where the global database is saved}

\item{video.description.folder}{directory with the video description file}

\item{video.description.file}{name of the video description file}

\item{total_frames}{length of video analysed in frames}
}
\value{
returns a data.table with the population densities, biovolume as well as the aggregated morphology and movement information for each population (sample)
}
\description{
Takes the data comprising the information for each frame and calculates summary statistics such as mean and sd (for all morphology metrics) and mean, sd and min/max 
for some of the movement metrics along the trajectory. Values are rounded to the second decimal.
}
\examples{
summarize_populations()
}
