% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_threshold_values.R
\name{check_threshold_values}
\alias{check_threshold_values}
\title{Function to assist finding good thresholds used for the segmentation of the difference video. If you run bemovi for the first time, verify with this function that all target objects are
properly identified
 
This function creates an ImageJ macro that can be helpful for checking the thresholds specified in the user section; the macro will be saved in the ImageJ macro directory in the working directory and
then automatically opened in ImageJ. Depending on the video size, it might take some time to open the video and the thresholding menu. The default thresholds values 
should be adjusted in the ImageJ macro, until appropriate values are found. These values should then be used in the R functions / script, e.g., when calling the Locate_and_measure_particles function.}
\usage{
check_threshold_values(to.data = par_to.data(),
  raw.video.folder = par_raw.video.folder(),
  ijmacs.folder = par_ijmacs.folder(), vid_select = 0,
  difference.lag = par_difference.lag(), thresholds = par_thresholds(),
  IJ.path = par_IJ.path(), memory = par_memory())
}
\arguments{
\item{to.data}{path to the working directory}

\item{raw.video.folder}{directory with the raw video files}

\item{ijmacs.folder}{directory where the check_trehsolds macro is saved}

\item{vid_select}{video selected to find appropriate thresholds; default is the first video}

\item{difference.lag}{numeric value specifying the offset between two frames of a video}

\item{thresholds}{Numeric vector containing the min and max threshold values}

\item{IJ.path}{path to ImageJ folder, containing the ij.jar executable}

\item{memory}{numeric value specifying the amount of memory available to ImageJ (defaults to 512)}
}
\description{
Function to assist finding good thresholds used for the segmentation of the difference video. If you run bemovi for the first time, verify with this function that all target objects are
properly identified
 
This function creates an ImageJ macro that can be helpful for checking the thresholds specified in the user section; the macro will be saved in the ImageJ macro directory in the working directory and
then automatically opened in ImageJ. Depending on the video size, it might take some time to open the video and the thresholding menu. The default thresholds values 
should be adjusted in the ImageJ macro, until appropriate values are found. These values should then be used in the R functions / script, e.g., when calling the Locate_and_measure_particles function.
}
