% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_overlay_videos.R
\name{create_overlays}
\alias{create_overlays}
\title{Function to create a new video with the extracted trajectories overlayed onto the original video}
\usage{
create_overlays(to.data = par_to.data(),
  merged.data.folder = par_merged.data.folder(),
  raw.video.folder = par_raw.video.folder(),
  temp.overlay.folder = par_temp.overlay.folder(),
  overlay.folder = par_overlay.folder(), width = par_width(),
  height = par_height(), difference.lag = par_difference.lag(),
  type = "traj", predict_spec = FALSE, contrast.enhancement = 0,
  IJ.path = par_IJ.path(), memory = par_memory())
}
\arguments{
\item{to.data}{path to the working directory}

\item{merged.data.folder}{directory where the global database is saved}

\item{raw.video.folder}{directory with the raw video files}

\item{temp.overlay.folder}{temporary directory to save the overlay created with R}

\item{overlay.folder}{directory where the overlay videos are saved}

\item{width}{width of the raw video}

\item{height}{height of the raw video}

\item{difference.lag}{numeric value specifying the offset between two video frames to compute the difference image}

\item{type}{string indicating the visualization type (i.e. 'label' or 'traj'): either the overlay
is showing the trajectory ID and outlines the detected particle (type='label') or the whole trajectory
remains plotted (type='traj').}

\item{predict_spec}{logical If TRUE, the Master.rds file must have a column called predict_spec, indicating the species to which the trajectory belongs}

\item{contrast.enhancement}{numeric value to increase the contrast of the original video}

\item{IJ.path}{path to ImageJ folder, containing the 'ij.jar' executable}

\item{memory}{numeric value specifying the amount of memory available to ImageJ (defaults to 512)}
}
\description{
A function to overlay the extracted trajectories onto the original video, using plots created in R and then processed in 
ImageJ; two visualization types are available
}
