% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_overlay_subtitles.R
\name{create_overlays_subtitle}
\alias{create_overlays_subtitle}
\title{Function to make overlays using the subtitle files,}
\usage{
create_overlays_subtitle(
  to.data = par_to.data(),
  merged.data.folder = par_merged.data.folder(),
  avi.video.folder = par_raw.video.folder(),
  temp.overlay.folder = par_temp.overlay.folder(),
  overlay.folder = par_overlay.folder(),
  difference.lag = par_difference.lag(),
  ffmpeg = "ffmpeg",
  master = par_master(),
  overlay.type = "both",
  mc.cores = par_mc.cores()
)
}
\arguments{
\item{to.data}{path to the working directory}

\item{temp.overlay.folder}{temporary directory to save the overlay subtitles (.ssa files)}

\item{overlay.folder}{directory where the overlay videos are saved}

\item{overlay.type}{option for the overlays. Overlays can either be shown as
trajectory numbers ("number"), circle ("circle") or both ("both"). Defaults to "both".}

\item{traj.data}{filtered trajectory file}

\item{raw.avi.folder}{path to the folder containing the converted and compressed .avi files}

\item{fps}{framerate in the original video}

\item{vid.length}{duration in seconds of the video}

\item{width}{width in pixels of the original videos}

\item{height}{height in pixels of the original video}

\item{tools.path}{path to the dependency folder}
}
\value{
returns nothing (NULL)
}
\description{
Function creates subtitle commands for every cell in every frame of the videos, using the x/y coordinates.
Then ffmpeg is called to burn the overlay subtitles on, and save a compressed video.
}
