% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_overlay_subtitles.R
\name{create_overlays_subtitle}
\alias{create_overlays_subtitle}
\title{Function to make overlays using the subtitle files,}
\usage{
create_overlays_subtitle(
  to.data = par_to.data(),
  merged.data.folder = par_merged.data.folder(),
  raw.video.folder = par_raw.video.folder(),
  temp.overlay.folder = par_temp.overlay.folder(),
  overlay.folder = par_overlay.folder(),
  ffmpeg = par_ffmpeg(),
  master = par_master(),
  overlay.type = "both",
  mc.cores = par_mc.cores()
)
}
\arguments{
\item{to.data}{path to the working directory}

\item{merged.data.folder}{directory where the global database is saved}

\item{temp.overlay.folder}{temporary directory to save the overlay subtitles (.ssa files)}

\item{overlay.folder}{directory where the overlay videos are saved}

\item{ffmpeg}{command to run ffmpeg. The default is \code{par_ffmpeg()}.  It can include a path.}

\item{master}{name of the master file. Defaults to \code{par_master()}}

\item{overlay.type}{option for the overlays. Overlays can either be shown as}

\item{mc.cores}{number of cores toi be used for parallel execution. Defaults to \code{par_mac.cores()}}

\item{raw.avi.folder}{path to the folder containing the converted and compressed .avi files}
}
\value{
returns invisibly \code{NULL}
}
\description{
Function creates subtitle commands for every cell in every frame of the videos, using the x/y coordinates.
Then ffmpeg is called to burn the overlay subtitles on, and save a compressed video.
}
