% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Function to filter trajectories based on movement and detection rate characteristics}
\usage{
filter_data(raw_data, net_filter, duration_filter, detect_filter,
  median_step_filter)
}
\arguments{
\item{raw_data}{dataframe containing the rawdata: X- and Y-coordinates, frame, file and trajectory name, morphology and movement metrics}

\item{net_filter}{minimum net displacement to be considered a valid trajectory (in the length scale specified)}

\item{duration_filter}{minimum duration to be considered a valid trajectory (in seconds)}

\item{detect_filter}{minimum detection rate to be considered a valid trajectory (a proportion between 0 and 1)}

\item{median_step_filter}{threshold such that half of the step lengths are above the specified value}
}
\value{
returns a dataset with all fixes of valid trajectories
}
\description{
The function creates a dataframe containing all trajectories that are valid for further analysis by selecting on minimum net displacement, detection rate, 
trajectory length and the median step_length
}
