% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{fRanco}
\alias{fRanco}
\alias{fRanco-package}
\title{fRanco - an R package to quantify abundance, traits and behaviour of organisms in videos}
\description{
fRanco provides functions to segment videos of organisms, identify individuals and their traits (morphology and movement) and
measure their abundances
}

