% Generated by roxygen2 (4.0.2): do not edit by hand
\name{locate_and_measure_particles}
\alias{locate_and_measure_particles}
\title{Function to extract morphological measurements and X- and Y-coordinates for moving particles}
\usage{
locate_and_measure_particles(to.data, raw.video.folder, particle.data.folder,
  difference.lag, min_size = 0, max_size = 10000, thresholds = c(10, 255),
  IJ.path, memory = 512)
}
\arguments{
\item{to.data}{path to the working directory}

\item{raw.video.folder}{directory with the raw video files}

\item{particle.data.folder}{directory to which the data is saved as a text file}

\item{difference.lag}{numeric value specifying the offset between two video frames to
compute the difference image}

\item{min_size}{minimum size for detection of particles}

\item{max_size}{maximum size for detection of particles}

\item{thresholds}{vector containing the min and max threshold values (defaults to c(10,255))}

\item{IJ.path}{path to ImageJ executable}

\item{memory}{numeric value specifying the amount of memory available to ImageJ (defaults to 512)}
}
\value{
saves the output of the ParticleAnalyzer function of ImageJ as a text file in the output directory and then assembles the data into a single database
called 'particle.RData'. This data.frame contains information about the following properties: the area (transversal cut),
the mean, minimum and maximum of the grey value, the perimeter, width, length and angle with the dominant-axis of a fitted ellipse, and finally shape parameters such as
circularity, aspect ratio, roundness and solidity. For details of the morphological output, please refer to http://rsbweb.nih.gov/ij/docs/guide/146-30.html
}
\description{
Function calls ImageJ software and its ParticleAnalyzer function to extract for each frame of the video
several morphological descriptors and the X- and Y-coordinates of all moving particles. All videos in the raw.video.folder are analyses, separately.
}

