% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Organise_link_data.R
\name{organise_link_data}
\alias{organise_link_data}
\title{Function to merge the output of the ParticleLinker into one large database}
\usage{
organise_link_data(to.data, trajectory.data.folder)
}
\arguments{
\item{to.data}{path to the working directory}

\item{trajectory.data.folder}{directory where the output of the ParticleLinker is saved}
}
\value{
saves the data containing the X- and Y coordinates of a given trajectory, the frame, the trajectory ID and the file name of the video
from which the data was extracted to disk
}
\description{
Merge the trajectory data from the ParticleLinker into one data file
}

