% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organise_particle_data.R
\name{organise_particle_data}
\alias{organise_particle_data}
\title{Function to merge the morphology and data on X- and Y-coordinates into one file for further processing}
\usage{
organise_particle_data(to.data = par_to.data(),
  particle.data.folder = par_particle.data.folder(),
  pixel_to_scale = par_pixel_to_scale())
}
\arguments{
\item{to.data}{path to the working directory}

\item{particle.data.folder}{directory to which the data is saved as a text file}

\item{pixel_to_scale}{TODO}
}
\description{
This function merges the files containing morphology and coordinates (one for each video) into large dataset,
and saves it to the directory where the single files are located
}
