% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Summarize_PopulationLevel.R
\name{summarize_populations}
\alias{summarize_populations}
\title{Function to extract population level parameters easily from BEMOVI output}
\usage{
summarize_populations(traj.data, sum.data, write = FALSE, to.data,
  merged.data.folder, video.description.folder, video.description.file)
}
\description{
The function will extract population-level information from trajectory.data
}
\author{
Emanuel Fronhofer
}

