% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarize_populations}
\alias{summarize_populations}
\title{Function to summarize the morphology, movement and their variability at the population (sample) level
adds population density estimates; does not care about different species}
\usage{
summarize_populations(traj.data, sum.data, write = FALSE, to.data,
  merged.data.folder, video.description.folder, video.description.file)
}
\arguments{
\item{write}{logical argument to indicate whether aggregated information should be saved to disk}

\item{to.data}{path to the working directory}

\item{merged.data.folder}{directory where the global database is saved}

\item{data}{dataframe with the information on morphology and movement for each frame}

\item{give.fps}{Give frames per second (in object named fps) if it is not already defined globally}
}
\value{
returns a data.table with the aggregated morphology and movement information for each trajectory
}
\description{
Takes the data comprising the information for each frame and calculates summary statistics such as mean and sd (for all morphology metrics) and mean, sd and min/max
for some of the movement metrics along the trajectory. Values are rounded to the second decimal.
}

