% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize_trajectories.R
\name{summarize_trajectories}
\alias{summarize_trajectories}
\title{Function to summarize the morphology, movement and their variability on the trajectory level}
\usage{
summarize_trajectories(
  data,
  calculate.median = T,
  write = FALSE,
  to.data,
  merged.data.folder,
  fps
)
}
\arguments{
\item{data}{dataframe with the information on morphology and movement for each frame}

\item{calculate.median}{logical value to indicate whether the median/IQR or the mean/SD summaries should be calculated for the morphology}

\item{write}{logical argument to indicate whether aggregated information should be saved to disk}

\item{to.data}{path to the working directory}

\item{merged.data.folder}{directory where the global database is saved}

\item{fps}{Frames Per Second of the video}
}
\value{
returns a data.table with the aggregated morphology and movement information for each trajectory
}
\description{
Takes the data comprising the information for each frame and calculates summary statistics such as mean and sd (for all morphology metrics) and mean, sd and min/max 
for some of the movement metrics along the trajectory. Values are rounded to the second decimal.
}
